agentcore-gateway-and-strands-agent/agent/requirements.txt
bedrock-agentcore
strands-agents
bedrock-agentcore-starter-toolkit
boto3